First commit of regex parser
Yeah, this is a huge mess and should be cleaned up. This initial stab worked
for about 90% of the testcases but some of the more important ones would fail
because we weren't able to handle multiple modifiers in a row.

What this code should do is make a pass through the regex, look for interesting
operators (replacing them with symbols or a proc returning true : false),
and then we read through the string going through the operators, short
circuiting if any of them fails