Rewrite regex implementation
Found some cases that didn't work because the approach for * was too greedy.
Basically, if you have .*d and you need to match abcd, then the * shouldn't
match the d (but you might need to) so you need to be able to backtrack
or iteratively try selecting more and more of the string.